Move journey in to the new journey package
