+++ src/journey/index.js
@@ -1,0 +1,4 @@
+const journey = require('./journey');
+const RequestBoundJourney = require('./RequestBoundJourney');
+
+module.exports = { journey, RequestBoundJourney };
